# api-docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TAQAT",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/workspace/packages/{package}": {
            "put": {
                "operationId": "package.update",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Package updated successfully."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "package.destroy",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Package deleted successfully."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/workspace/packages/{package}/unassign": {
            "put": {
                "operationId": "package.unassign",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignPackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PackageResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Package unassigned from member."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/workspace/photos": {
            "post": {
                "operationId": "photo.store",
                "summary": "POST /api/workspace/photos \u2014 append photos to the owner's workspace",
                "tags": [
                    "Photo"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PhotoUploadRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/WorkspaceResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Photos uploaded."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "This action is unauthorized."
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "You have no workspace."
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "photo.destroy",
                "summary": "DELETE /api/workspace/photos \u2014 remove one photo by its stored path",
                "tags": [
                    "Photo"
                ],
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/WorkspaceResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Photo removed."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "errors": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "message",
                                                "errors"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "const": "You have no workspace."
                                                },
                                                "errors": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "message",
                                                "errors"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "This action is unauthorized."
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/seats/{seat}": {
            "put": {
                "operationId": "seat.update",
                "tags": [
                    "Seat"
                ],
                "parameters": [
                    {
                        "name": "seat",
                        "in": "path",
                        "required": true,
                        "description": "The seat ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSeatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SeatResource"
                                        },
                                        "message": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "seat.destroy",
                "tags": [
                    "Seat"
                ],
                "parameters": [
                    {
                        "name": "seat",
                        "in": "path",
                        "required": true,
                        "description": "The seat ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Seat deleted."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/seats/{seat}/unassign": {
            "put": {
                "operationId": "seat.unassign",
                "tags": [
                    "Seat"
                ],
                "parameters": [
                    {
                        "name": "seat",
                        "in": "path",
                        "required": true,
                        "description": "The seat ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SeatResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Seat unassigned."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AssignPackageRequest": {
                "type": "object",
                "properties": {
                    "member_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "member_id"
                ],
                "title": "AssignPackageRequest"
            },
...
            "PackageResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "workspace_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "speed_mbps": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "string"
                    },
                    "data_limit_gb": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "is_unlimited": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "assigned_members_count": {
                        "type": "integer"
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "avatar": {
                                    "type": [
                                        "string",
                                        "null"
                                    ]
                                },
                                "assigned_at": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "id",
                                "name",
                                "avatar",
                                "assigned_at"
                            ]
                        }
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "workspace_id",
                    "name",
                    "speed_mbps",
                    "price",
                    "data_limit_gb",
                    "is_unlimited",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "PackageResource"
            },
            "PhotoUploadRequest": {
                "type": "object",
                "properties": {
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary",
                            "contentMediaType": "application/octet-stream",
                            "maxLength": 5120
                        },
                        "minItems": 1,
                        "maxItems": 10
                    }
                },
                "required": [
                    "photos"
                ],
                "title": "PhotoUploadRequest"
            },
...
            "UpdatePackageRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 120
                    },
                    "speed_mbps": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0
                    },
                    "data_limit_gb": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1
                    },
                    "is_unlimited": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "title": "UpdatePackageRequest"
            },
...
            "UpdateSeatRequest": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "fixed",
                            "flexible",
                            "private_office"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "occupied",
                            "reserved",
                            "maintenance"
                        ]
                    },
                    "notes": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 500
                    }
                },
                "title": "UpdateSeatRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}